Separate BlockMaker baseclass and bitcoin implementation kafkaConsumerSolvedShare initialize on baseclass remove init on bytom and eth
