+++ src/BlockMaker.h
@@ -29,13 +29,7 @@
 #include "MySQLConnection.h"
 #include "Stratum.h"
 
-#include <uint256.h>
-
-#include <deque>
 #include <vector>
-#include <boost/date_time/posix_time/posix_time.hpp>
-
-namespace bpt = boost::posix_time;
 
 struct NodeDefinition
 {
@@ -51,53 +45,6 @@ struct BlockMakerDefinition
   string solvedShareTopic_;
 };
 
-// class BlockMakerHandler {
-//   public:
-//     virtual ~BlockMakerHandler() = 0; // mark it's an abstract class
-//     virtual void init(const BlockMakerDefinition &def) { def_ = def; }
-    
-//     // read-only definition
-//     virtual const BlockMakerDefinition& def() { return def_; }
-//     virtual void processSolvedShare(rd_kafka_message_t *rkmessage) = 0;
-//     // Interface with the GwMaker.
-//     // There is a default implementation that use virtual functions below.
-//     // If the implementation does not meet the requirements, you can overload it
-//     // and ignore all the following virtual functions.
-//   //   virtual string makeRawGwMsg();
-
-//   // protected:
-
-//   //   // These virtual functions make it easier to implement the makeRawGwMsg() interface.
-//   //   // In most cases, you just need to override getRequestData() and processRawGw().
-//   //   // If you have overloaded makeRawGwMsg() above, you can ignore all the following functions.
-
-//   //   // Receive rpc response and generate RawGw message for the pool.
-//   //   virtual string processRawGw(const string &gw) { return ""; }
-
-//   //   // Call RPC `getwork` and get the response.
-//   //   virtual bool callRpcGw(string &resp);
-
-//   //   // Body of HTTP POST used by callRpcGw().
-//   //   // return "" if use HTTP GET.
-//   //   virtual string getRequestData() { return ""; }
-//   //   // HTTP header `User-Agent` used by callRpcGw().
-//   //   virtual string getUserAgent() { return "curl"; }
-
-//     // blockchain and RPC-server definitions
-//     BlockMakerDefinition def_;
-// };
-
-// class BlockMakerHandlerEth : public BlockMakerHandler{
-//   virtual void processSolvedShare(rd_kafka_message_t *rkmessage);
-// };
-
-// class BlockMakerHandlerSia : public BlockMakerHandler{
-//   virtual void processSolvedShare(rd_kafka_message_t *rkmessage) {
-    
-//   }
-// };
-
-
 ////////////////////////////////// BlockMaker //////////////////////////////////
 class BlockMaker
 {
@@ -105,26 +52,7 @@ protected:
   BlockMakerDefinition def_;
   atomic<bool> running_;
 
-  mutex rawGbtLock_;
-  size_t kMaxRawGbtNum_; // how many rawgbt should we keep
-  // key: gbthash
-  std::deque<uint256> rawGbtQ_;
-  // key: gbthash, value: block template json
-  std::map<uint256, shared_ptr<vector<CTransactionRef>>> rawGbtMap_;
-
-  mutex jobIdMapLock_;
-  size_t kMaxStratumJobNum_;
-  // key: jobId, value: gbthash
-  std::map<uint64_t, uint256> jobId2GbtHash_;
-
-  bpt::ptime lastSubmittedBlockTime;
-  uint32_t submittedRskBlocks;
-
-  KafkaConsumer kafkaConsumerRawGbt_;
-  KafkaConsumer kafkaConsumerStratumJob_;
   KafkaConsumer kafkaConsumerSolvedShare_;
-  KafkaConsumer kafkaConsumerNamecoinSolvedShare_;
-  KafkaConsumer kafkaConsumerRskSolvedShare_;
 
   // submit new block to bitcoind
   // pair: <RpcAddress, RpcUserpass>
@@ -132,67 +60,9 @@ protected:
 
   MysqlConnectInfo poolDB_; // save blocks to table.found_blocks
 
-  void insertRawGbt(const uint256 &gbtHash,
-                    shared_ptr<vector<CTransactionRef>> vtxs);
-
-  thread threadConsumeRawGbt_;
-  thread threadConsumeStratumJob_;
-  thread threadConsumeNamecoinSolvedShare_;
-  thread threadConsumeRskSolvedShare_;
-
-  void runThreadConsumeRawGbt();
   void runThreadConsumeSolvedShare();
-  void runThreadConsumeStratumJob();
-  void runThreadConsumeNamecoinSolvedShare();
-  void runThreadConsumeRskSolvedShare();
-
-  void consumeRawGbt(rd_kafka_message_t *rkmessage);
-  void consumeStratumJob(rd_kafka_message_t *rkmessage);
   void consumeSolvedShare(rd_kafka_message_t *rkmessage);
-  virtual void processSolvedShare(rd_kafka_message_t *rkmessage);
-  void consumeNamecoinSolvedShare(rd_kafka_message_t *rkmessage);
-  void consumeRskSolvedShare(rd_kafka_message_t *rkmessage);
-
-  void addRawgbt(const char *str, size_t len);
-
-  void saveBlockToDBNonBlocking(const FoundBlock &foundBlock,
-                                const CBlockHeader &header,
-                                const uint64_t coinbaseValue, const int32_t blksize);
-  void _saveBlockToDBThread(const FoundBlock &foundBlock,
-                            const CBlockHeader &header,
-                            const uint64_t coinbaseValue, const int32_t blksize);
-
-  void submitBlockNonBlocking(const string &blockHex);
-  void _submitBlockThread(const string &rpcAddress, const string &rpcUserpass,
-                          const string &blockHex);
-  bool checkBitcoinds();
-
-  void submitNamecoinBlockNonBlocking(const string &auxBlockHash,
-                                      const string &auxPow,
-                                      const string &bitcoinBlockHash,
-                                      const string &rpcAddress,
-                                      const string &rpcUserpass);
-  void _submitNamecoinBlockThread(const string &auxBlockHash,
-                                  const string &auxPow,
-                                  const string &bitcoinBlockHash,
-                                  const string &rpcAddress,
-                                  const string &rpcUserpass);
-
-  void submitRskBlockPartialMerkleNonBlocking(const string &rpcAddress,
-                                              const string &rpcUserPwd,
-                                              const string &blockHashHex, 
-                                              const string &blockHeaderHex, 
-                                              const string &coinbaseHex, 
-                                              const string &merkleHashesHex, 
-                                              const string &totalTxCount);
-  void _submitRskBlockPartialMerkleThread(const string &rpcAddress,
-                                          const string &rpcUserPwd,
-                                          const string &blockHashHex, 
-                                          const string &blockHeaderHex, 
-                                          const string &coinbaseHex, 
-                                          const string &merkleHashesHex, 
-                                          const string &totalTxCount);
-  bool submitToRskNode();
+  virtual void processSolvedShare(rd_kafka_message_t *rkmessage) = 0;
 
 public:
   BlockMaker(const BlockMakerDefinition& def, const char *kafkaBrokers, const MysqlConnectInfo &poolDB);
@@ -201,8 +71,8 @@ public:
   void addNodeRpc(const string &rpcAddress, const string &rpcUserpass);
 
   virtual bool init();
-  void stop();
-  void run();
+  virtual void stop();
+  virtual void run();
 };
