+++ src/Utils.cc
@@ -290,7 +290,7 @@ bool httpPOST(const char *url, const char *userpwd, const char *postData,
   return httpPOST(url, userpwd, postData, response, timeoutMs, mineType, "curl");
 }
 
-bool bitcoindRpcCall(const char *url, const char *userpwd, const char *reqData,
+bool blockchainNodeRpcCall(const char *url, const char *userpwd, const char *reqData,
                      string &response) {
   return httpPOST(url, userpwd, reqData, response, 5000/* timeout ms */, "application/json");
 }
@@ -412,60 +412,6 @@ string getStatsFilePath(const char *chainType, const string &dataDir, time_t ts)
                          dataDir.c_str(), needSlash ? "/" : "",
                          chainType,
                          date("%F", ts).c_str());
-}
-
-bool checkBitcoinRPC(const string &rpcAddr, const string &rpcUserpass) {
-  string response;
-  string request = "{\"jsonrpc\":\"1.0\",\"id\":\"1\",\"method\":\"getnetworkinfo\",\"params\":[]}";
-  bool res = bitcoindRpcCall(rpcAddr.c_str(), rpcUserpass.c_str(),
-                             request.c_str(), response);
-  if (!res) {
-    LOG(ERROR) << "rpc call failure";
-    return false;
-  }
-
-  LOG(INFO) << "getnetworkinfo: " << response;
-
-  JsonNode r;
-  if (!JsonNode::parse(response.c_str(),
-                       response.c_str() + response.length(), r)) {
-    LOG(ERROR) << "decode getnetworkinfo failure";
-    return false;
-  }
-
-  // check if the method not found
-  if (r["result"].type() != Utilities::JS::type::Obj) {
-    LOG(INFO) << "node doesn't support getnetworkinfo, try getinfo";
-
-    request = "{\"jsonrpc\":\"1.0\",\"id\":\"1\",\"method\":\"getinfo\",\"params\":[]}";
-    res = bitcoindRpcCall(rpcAddr.c_str(), rpcUserpass.c_str(),
-                          request.c_str(), response);
-    if (!res) {
-      LOG(ERROR) << "rpc call failure";
-      return false;
-    }
-
-    LOG(INFO) << "getinfo: " << response;
-
-    if (!JsonNode::parse(response.c_str(),
-                         response.c_str() + response.length(), r)) {
-      LOG(ERROR) << "decode getinfo failure";
-      return false;
-    }
-  }
-
-  // check fields & connections
-  if (r["result"].type() != Utilities::JS::type::Obj ||
-      r["result"]["connections"].type() != Utilities::JS::type::Int) {
-    LOG(ERROR) << "getnetworkinfo missing some fields";
-    return false;
-  }
-  if (r["result"]["connections"].int32() <= 0) {
-    LOG(ERROR) << "node connections is zero";
-    return false;
-  }
-
-  return true;
 }
 
 // A 37-character character set.
