+++ src/bytom/BlockMakerBytom.h
@@ -32,7 +32,6 @@ class BlockMakerBytom : public BlockMaker
 public:
   BlockMakerBytom(const BlockMakerDefinition &def, const char *kafkaBrokers, const MysqlConnectInfo &poolDB);
   void processSolvedShare(rd_kafka_message_t *rkmessage) override;
-  bool init() override;
 
 private:
   void submitBlockNonBlocking(const string &request);
