+++ src/eth/BlockMakerEth.h
@@ -33,7 +33,6 @@ class BlockMakerEth : public BlockMaker
 public:
   BlockMakerEth(const BlockMakerDefinition& def, const char *kafkaBrokers, const MysqlConnectInfo &poolDB);
   void processSolvedShare(rd_kafka_message_t *rkmessage) override;
-  bool init() override;
 
 private:
   void submitBlockNonBlocking(const string &blockJson);
